requirements.txt
openai==0.28.0
Mastodon.py>=1.7.1
python-dotenv>=1.0.0
pytest>=7.0.0
